+++ R/gg_line.R
@@ -277,6 +277,7 @@ gg_line <- function(data,
 #' @param col_var Unquoted categorical variable for lines and points to be coloured by. Required input.
 #' @param text_var Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.
 #' @param pal Character vector of hex codes. 
+#' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param size_point Size of points. Defaults to 1. 
 #' @param size_line Size of lines. Defaults to 0.5. 
@@ -338,6 +339,7 @@ gg_line_col <- function(data,
                         col_var,
                         text_var = NULL,
                         pal = NULL,
+                        pal_na = "#7F7F7FFF",
                         pal_rev = FALSE,
                         size_point = 1,
                         size_line = 0.5,
@@ -559,7 +561,7 @@ gg_line_col <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = pal_na(),
+      na.value = pal_na,
       name = stringr::str_wrap(col_title, col_title_wrap)
     )
   
@@ -888,6 +890,7 @@ gg_line_facet <- function(data,
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param text_var Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.
 #' @param pal Character vector of hex codes. 
+#' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param size_point Size of points. Defaults to 1. 
 #' @param size_line Size of lines. Defaults to 0.5. 
@@ -955,6 +958,7 @@ gg_line_col_facet <- function(data,
                               facet_var,
                               text_var = NULL,
                               pal = NULL,
+                              pal_na = "#7F7F7FFF",
                               pal_rev = FALSE,
                               size_point = 1,
                               size_line = 0.5,
@@ -1198,7 +1202,7 @@ gg_line_col_facet <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = pal_na(),
+      na.value = pal_na,
       name = stringr::str_wrap(col_title, col_title_wrap)
     ) +
     labs(