+++ R/gg_point.R
@@ -282,6 +282,7 @@ gg_point <- function(data,
 #' @param alpha The opacity of points. Defaults to 1.
 #' @param size_point Size of points. Defaults to 1.
 #' @param pal Character vector of hex codes. 
+#' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param title Title string. Defaults to NULL.
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 100. Not applicable where mobile equals TRUE.
@@ -312,10 +313,10 @@ gg_point <- function(data,
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.  
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and an internal function for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.  
-#' @param col_labels_dp For numeric colour methods, the number of decimal places of numeric labels. Defaults to the maximum.    
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
 #' @param col_pretty_n For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 4. 
+#' @param col_right_closed For a numeric colour variable, TRUE or FALSE of whether bins or quantiles are to be cut right-closed. Defaults to TRUE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. Not applicable where mobile equals TRUE.
 #' @param caption Caption title string. 
@@ -344,6 +345,7 @@ gg_point_col <- function(data,
                          alpha = 1,
                          size_point = 1,
                          pal = NULL,
+                         pal_na = "#7F7F7FFF",
                          pal_rev = FALSE,
                          x_balance = FALSE,
                          x_expand = NULL,
@@ -372,10 +374,10 @@ gg_point_col <- function(data,
                          caption = NULL,
                          col_cuts = NULL,
                          col_labels = NULL,
-                         col_labels_dp = NULL,
                          col_method = NULL,
                          col_na = TRUE,
                          col_pretty_n = 4,
+                         col_right_closed = TRUE,
                          font_family = "",
                          font_size_title = NULL,
                          font_size_body = NULL,
@@ -468,11 +470,17 @@ gg_point_col <- function(data,
       })
     }
 
-    data <- data %>% 
-      dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
+    if (is.null(col_labels)) col_labels <- scales::comma
     
-    if (is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
-    if (is.null(col_labels)) col_labels <- sv_cuts_to_labels(col_cuts, col_labels_dp)
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!col_var, ~kimisc::cut_format(.x, col_cuts, 
+                                                                 right = col_right_closed, 
+                                                                 include.lowest = TRUE, 
+                                                                 dig.lab = 50, 
+                                                                 ordered_result = TRUE,
+                                                                 format_fun = col_labels)))
+    
+    col_labels <- sv_label_intervals
     
     col_n <- length(col_cuts) - 1
     if (is.null(pal)) pal <- pal_viridis_reorder(col_n)
@@ -606,7 +614,7 @@ gg_point_col <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = pal_na(), 
+      na.value = pal_na, 
       name = stringr::str_wrap(col_title, col_title_wrap)
     ) 
   
@@ -940,6 +948,7 @@ gg_point_facet <- function(data,
 #' @param alpha The opacity of points. Defaults to 1.
 #' @param size_point Size of points. Defaults to 1.
 #' @param pal Character vector of hex codes. 
+#' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param title Title string. Defaults to NULL.
 #' @param title_wrap Number of characters to wrap the title to. Defaults to 100. 
@@ -970,10 +979,10 @@ gg_point_facet <- function(data,
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.  
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and an internal function for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.  
-#' @param col_labels_dp For numeric colour methods, the number of decimal places of numeric labels. Defaults to the maximum.    
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
 #' @param col_pretty_n For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 4. 
+#' @param col_right_closed For a numeric colour variable, TRUE or FALSE of whether bins or quantiles are to be cut right-closed. Defaults to TRUE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. 
 #' @param facet_labels A function or named vector to modify facet scale labels. Defaults to converting labels to sentence case. Use ggplot2::waiver() to keep facet labels untransformed.
@@ -1009,6 +1018,7 @@ gg_point_col_facet <-
            alpha = 1,
            size_point = 1,
            pal = NULL,
+           pal_na = "#7F7F7FFF",
            pal_rev = FALSE,
            title = NULL,
            title_wrap = 100,
@@ -1039,10 +1049,10 @@ gg_point_col_facet <-
            y_zero_line = NULL,
            col_cuts = NULL,
            col_labels = NULL,
-           col_labels_dp = NULL,
            col_method = NULL,
            col_na = TRUE,
            col_pretty_n = 4,
+           col_right_closed = TRUE,
            col_title = NULL,
            col_title_wrap = 25,
            facet_labels = stringr::str_to_sentence,
@@ -1150,12 +1160,18 @@ gg_point_col_facet <-
         })
       }
 
-      data <- data %>% 
-        dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
+      if (is.null(col_labels)) col_labels <- scales::comma
       
-      if (is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
-      if (is.null(col_labels)) col_labels <- sv_cuts_to_labels(col_cuts, col_labels_dp)
-
+      data <- data %>% 
+        dplyr::mutate(dplyr::across(!!col_var, ~kimisc::cut_format(.x, col_cuts, 
+                                                                   right = col_right_closed, 
+                                                                   include.lowest = TRUE, 
+                                                                   dig.lab = 50, 
+                                                                   ordered_result = TRUE,
+                                                                   format_fun = col_labels)))
+      
+      col_labels <- sv_label_intervals
+      
       col_n <- length(col_cuts) - 1
       if (is.null(pal)) pal <- pal_viridis_reorder(col_n)
       else pal <- pal[1:col_n]
@@ -1290,7 +1306,7 @@ gg_point_col_facet <-
         values = pal,
         drop = FALSE,
         labels = col_labels,
-        na.value = pal_na(),
+        na.value = pal_na,
         name = stringr::str_wrap(col_title, col_title_wrap)
       ) +
       labs(