+++ R/gg_sf.R
@@ -137,6 +137,7 @@ gg_sf <- function(data,
 #' @param col_var Unquoted variable for points to be coloured by. Required input.
 #' @param text_var Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.
 #' @param pal Character vector of hex codes. Defaults to NULL, which selects the colorbrewer Set1 or viridis.
+#' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param size_point Size of points. Defaults to 0.5.
 #' @param size_line Size of lines. Defaults to 0.5.
@@ -151,10 +152,10 @@ gg_sf <- function(data,
 #' @param subtitle_wrap Number of characters to wrap the subtitle to. Defaults to 100. Not applicable where mobile equals TRUE.
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and an internal function for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.  
-#' @param col_labels_dp For numeric colour methods, the number of decimal places of numeric labels. Defaults to the maximum.    
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_pretty_n For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 4. 
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_right_closed For a numeric colour variable, TRUE or FALSE of whether bins or quantiles are to be cut right-closed. Defaults to TRUE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. Not applicable where mobile equals TRUE.
 #' @param caption Caption title string. 
@@ -186,6 +187,7 @@ gg_sf_col <- function(data,
                       col_var,
                       text_var = NULL,
                       pal = NULL,
+                      pal_na = "#7F7F7FFF",
                       pal_rev = FALSE,
                       size_point = 1,
                       size_line = 0.5,
@@ -200,10 +202,10 @@ gg_sf_col <- function(data,
                       subtitle_wrap = 100,
                       col_cuts = NULL,
                       col_labels = NULL,
-                      col_labels_dp = NULL,
                       col_na = TRUE,
                       col_pretty_n = 4,
                       col_method = NULL,
+                      col_right_closed = TRUE,
                       col_title = NULL,
                       col_title_wrap = 25,
                       caption = NULL,
@@ -285,11 +287,17 @@ gg_sf_col <- function(data,
       })
     }
 
-    data <- data %>% 
-      dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
+    if (is.null(col_labels)) col_labels <- scales::comma
     
-    if (is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
-    if (is.null(col_labels)) col_labels <- sv_cuts_to_labels(col_cuts, col_labels_dp)
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!col_var, ~kimisc::cut_format(.x, col_cuts, 
+                                                                 right = col_right_closed, 
+                                                                 include.lowest = TRUE, 
+                                                                 dig.lab = 50, 
+                                                                 ordered_result = TRUE,
+                                                                 format_fun = col_labels)))
+    
+    col_labels <- sv_label_intervals
     
     col_n <- length(col_cuts) - 1
     if (is.null(pal)) pal <- pal_viridis_reorder(col_n)
@@ -342,7 +350,7 @@ gg_sf_col <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = pal_na(),
+      na.value = pal_na,
       name = stringr::str_wrap(col_title, col_title_wrap)
     )
   
@@ -352,7 +360,7 @@ gg_sf_col <- function(data,
         values = pal,
         drop = FALSE,
         labels = col_labels,
-        na.value = pal_na(),
+        na.value = pal_na,
         name = stringr::str_wrap(col_title, col_title_wrap)
       )
   }
@@ -564,6 +572,7 @@ gg_sf_facet <- function(data,
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param text_var Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.
 #' @param pal Character vector of hex codes. Defaults to NULL, which selects the colorbrewer Set1 or viridis.
+#' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
 #' @param size_point Size of points. Defaults to 0.5.
 #' @param size_line Size of lines. Defaults to 0.5.
@@ -579,10 +588,10 @@ gg_sf_facet <- function(data,
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
 #' @param facet_labels A function or named vector to modify facet scale labels. Defaults to converting labels to sentence case. Use ggplot2::waiver() to keep facet labels untransformed.
 #' @param col_labels A function or named vector to modify colour scale labels. Defaults to stringr::str_to_sentence for categorical colour variables and an internal function for numeric colour variables. Use ggplot2::waiver() to keep colour labels untransformed.  
-#' @param col_labels_dp For numeric colour methods, the number of decimal places of numeric labels. Defaults to the maximum.    
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "bin".
 #' @param col_pretty_n For a numeric colour variable of "bin" col_method, the desired number of intervals on the colour scale, as calculated by the pretty algorithm. Defaults to 4. 
 #' @param col_na TRUE or FALSE of whether to include col_var NA values. Defaults to TRUE.
+#' @param col_right_closed For a numeric colour variable, TRUE or FALSE of whether bins or quantiles are to be cut right-closed. Defaults to TRUE.
 #' @param col_title Colour title string for the legend. Defaults to NULL, which converts to sentence case with spaces. Use "" if you would like no title.
 #' @param col_title_wrap Number of characters to wrap the colour title to. Defaults to 25. 
 #' @param facet_na TRUE or FALSE of whether to include facet_var NA values. Defaults to TRUE.
@@ -606,6 +615,7 @@ gg_sf_col_facet <- function(data,
                             facet_var,
                             text_var = NULL,
                             pal = NULL,
+                            pal_na = "#7F7F7FFF",
                             pal_rev = FALSE,
                             size_point = 1,
                             size_line = 0.5,
@@ -620,10 +630,10 @@ gg_sf_col_facet <- function(data,
                             subtitle_wrap = 100,
                             col_cuts = NULL,
                             col_labels = NULL,
-                            col_labels_dp = NULL,
                             col_method = NULL,
                             col_na = TRUE,
                             col_pretty_n = 4,
+                            col_right_closed = TRUE,
                             col_title = NULL,
                             col_title_wrap = 25,
                             facet_labels = stringr::str_to_sentence,
@@ -721,11 +731,17 @@ gg_sf_col_facet <- function(data,
       })
     }
 
-    data <- data %>% 
-      dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
+    if (is.null(col_labels)) col_labels <- scales::comma
     
-    if (is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
-    if (is.null(col_labels)) col_labels <- sv_cuts_to_labels(col_cuts, col_labels_dp)
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!col_var, ~kimisc::cut_format(.x, col_cuts, 
+                                                                 right = col_right_closed, 
+                                                                 include.lowest = TRUE, 
+                                                                 dig.lab = 50, 
+                                                                 ordered_result = TRUE,
+                                                                 format_fun = col_labels)))
+    
+    col_labels <- sv_label_intervals
     
     col_n <- length(col_cuts) - 1
     if (is.null(pal)) pal <- pal_viridis_reorder(col_n)
@@ -776,7 +792,7 @@ gg_sf_col_facet <- function(data,
       values = pal,
       drop = FALSE,
       labels = col_labels,
-      na.value = pal_na(),
+      na.value = pal_na,
       name = stringr::str_wrap(col_title, col_title_wrap)
     )
   
@@ -786,7 +802,7 @@ gg_sf_col_facet <- function(data,
         values = pal,
         drop = FALSE,
         labels = col_labels,
-        na.value = pal_na(),
+        na.value = pal_na,
         name = stringr::str_wrap(col_title, col_title_wrap)
       )
   }